+++ app/constants/routes.ts
@@ -1,0 +1,8 @@
+/**
+ * 各ページへのルートを管理する定数
+ */
+export const ROUTES = {
+  ROOT: "/",
+  LOGIN: "/login",
+  LOGOUT: "/logout",
+} as const;
