+++ app/services/authService.ts
@@ -1,17 +1,6 @@
-import { sessionStorage } from "./session.server";
+import { sessionStorage } from "~/utils/server/session.server";
 import { redirect } from "@remix-run/node";
-
-/**
- * userInfo取得用関数
- */
-export const getUserSession = async (request: Request) => {
-  const session = await sessionStorage.getSession(
-    request.headers.get("Cookie")
-  );
-  const userInfo = session.get("userInfo");
-
-  return { session, userInfo };
-};
+import { ROUTES } from "~/constants/routes";
 
 /**
  * ログアウト用関数
@@ -23,7 +12,7 @@ export const logoutUser = async (request: Request) => {
   session.unset("userInfo");
   session.unset("accessToken");
 
-  return redirect("/login", {
+  return redirect(ROUTES.LOGIN, {
     headers: {
       "Set-Cookie": await sessionStorage.commitSession(session),
     },
