--- app/utils/session.server.ts
@@ -1,15 +1,0 @@
-import { createCookieSessionStorage } from "@remix-run/node";
-
-const sessionSecret = process.env.SESSION_SECRET || "";
-
-export const sessionStorage = createCookieSessionStorage({
-  cookie: {
-    name: "remix_session",
-    secure: process.env.NODE_ENV === "production",
-    maxAge: 60 * 60 * 24,
-    httpOnly: true,
-    path: "/",
-    sameSite: "lax",
-    secrets: [sessionSecret],
-  },
-});
